Allow upload of plugin configuration on the database #23
Added a file Web/js/plugins-controller.js to control the event of the
page index.html (e.g. form submit)
Added a file Web/api/upload-file.py , a flask server to handle the reception
of file uploaded though the web application

Requires the use of a web server apache configured with reversed proxy
to redirect the request from localhost:80/ajax/upload to
localhost:5000/upload (the port reserved by the flask server).

Requires that the Flask server is started and listening and that all
the module imported in the head of the file are installed on the filesystem.